Climber fixes

+++ src/main/java/frc/robot/Constants.java
@@ -22,10 +22,10 @@ public final class Constants {
     public static class Climber {
         public static final int TICKS_PER_METER = 0;
 
-        public static final int KP = 0;
-        public static final int KI = 0;
-        public static final int KD = 0;
-        public static final int KF = 0;
+        public static final double KP = 0;
+        public static final double KI = 0;
+        public static final double KD = 0;
+        public static final double KF = 0;
 
         public static final int CRUISE_VELOCITY = 0; //sensor units per 100ms.
         public static final int ACCELERATION = 0; //sensor units per 100ms^2.
